Use events to send responses to main process as they are returned

+++ src/index.ts
@@ -1,20 +1,31 @@
 import { PocketUtils } from './pocket-utils';
 import { isError, isString } from 'lodash';
 import { getRandom, getRandomInt, logError } from './util';
-import { RelayRequest, RelayResponse } from './interfaces';
+import { ProcessMessage, RelayRequest, RelayResponse } from './interfaces';
 import { Runner } from './runner';
 import { fork } from 'child_process';
 import path from 'path';
+import { messageEvent } from './constants';
+import colors from 'colors/safe';
+import { stdout as log } from 'single-line-log';
 
-const startRunner = (endpoint: string, chainId: string, length: number, requests: RelayRequest[]): Promise<RelayResponse[]> => {
+const startRunner = (endpoint: string, chainId: string, length: number, requests: RelayRequest[], onResponse: (response: RelayResponse)=>void): Promise<RelayResponse[]> => {
   return new Promise((resolve, reject) => {
     const instance = fork(path.join(__dirname, 'start-runner'));
     instance.on('message', message => {
-      resolve(message as RelayResponse[]);
-      instance.kill();
+      const { event, payload } = message as ProcessMessage;
+      if(event === messageEvent.RESPONSES) {
+        resolve(payload as RelayResponse[]);
+        instance.kill();
+      } else if(event === messageEvent.RESPONSE) {
+        onResponse(payload as RelayResponse);
+      }
     });
     instance.on('error', reject);
-    instance.send({endpoint, chainId, length, requests});
+    instance.send({
+      event: messageEvent.REQUESTS,
+      payload: {endpoint, chainId, length, requests}
+    });
   });
 };
 
@@ -33,8 +44,8 @@ const start = async function() {
 
   const chainId = '0021';
   const length = 1;
-  const totalRequests = 1200;
-  const instances = 2;
+  const totalRequests = 60;
+  const instances = 1;
 
   let startingBlock: number;
 
@@ -79,7 +90,8 @@ const start = async function() {
       data: JSON.stringify({
         id: getRandom(),
         jsonrpc: '2.0',
-        method: 'eth_blockNumber',
+        // method: 'eth_blockNumber',
+        method: 'net_version',
         params: [],
         // method: 'eth_getBlockByNumber',
         // params: ['0x' + (startingBlock - offset).toString(16), false],
@@ -90,18 +102,33 @@ const start = async function() {
     });
   }
 
+  const allResponses: RelayResponse[] = [];
+
+  let percentComplete = -1;
+  const onResponse = (response: RelayResponse) => {
+    allResponses.push(response);
+    const percent = Math.floor((allResponses.length / totalRequests) * 100);
+    if(percent !== percentComplete) {
+      let percentStr = percent.toString(10);
+      while(percentStr.length < 3) {
+        percentStr = ' ' + percentStr;
+      }
+      percentComplete = percent;
+      log(`${percentStr}% ${colors.grey('complete')}`);
+    }
+  };
+
   const totalRequestsForRunner = totalRequests / instances;
   const runners = [];
   for(let i = 0; i < instances; i++) {
     const start = i * totalRequestsForRunner;
     const runnerRequests = allRequests.slice(start, start + totalRequestsForRunner);
-    runners.push(() => startRunner(endpoint, chainId, length, runnerRequests));
+    runners.push(() => startRunner(endpoint, chainId, length, runnerRequests, onResponse));
   }
 
-  const allResponses: RelayResponse[][] = await Promise.all(runners.map(r => r()));
+  await Promise.all(runners.map(r => r()));
 
   const sortedResponses: RelayResponse[] = allResponses
-    .reduce((joined, arr) => [...joined, ...arr])
     .sort((a, b) => a.timestamp === b.timestamp ? 0 : a.timestamp > b.timestamp ? 1 : -1);
 
   // console.log(sortedResponses[0]);
@@ -110,13 +137,17 @@ const start = async function() {
 
   const averageDuration = sortedResponses.reduce((sum, r) => sum + r.duration, 0) / sortedResponses.length;
 
-  console.log(`\n reqs per second:  ${totalRequestsForRunner / (length * 60)}`);
-  console.log(`average duration:  ${averageDuration}`);
-  console.log(`  total requests:  ${sortedResponses.length}`);
-  console.log(`   success count:  ${sortedResponses.filter(r => !r.error).length}`);
-  console.log(`     error count:  ${sortedResponses.filter(r => !!r.error).length}`);
+  const successResponses = sortedResponses.filter(r => !r.error);
+  const errorResponses = sortedResponses.filter(r => !!r.error);
 
-  console.log('\nDone!');
+  const durationColor = averageDuration < 100 ? colors.green : averageDuration < 1000 ? colors.yellow : colors.red;
+
+  console.log('\n');
+  console.log(`average duration:  ${durationColor(averageDuration.toFixed(3))}`);
+  console.log(` reqs per second:  ${instances * totalRequestsForRunner / (length * 60)}`);
+  console.log(`  total requests:  ${sortedResponses.length}`);
+  console.log(`   success count:  ${colors.green(successResponses.length.toString(10))}`);
+  console.log(`     error count:  ${errorResponses.length > 0 ? colors.red(errorResponses.length.toString(10)) : errorResponses.length.toString(10)}`);
 
 };
